app/api/deps.py
from typing import Generator, Optional
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from jose import jwt
from pydantic import ValidationError
from beanie import PydanticObjectId

from app.core.config import settings
from app.models.user import User
from app.schemas.user import UserInDB

# This will be used for dependency injection in protected routes
def get_current_user():
    # For now, we'll return a mock user
    # In a real application, you would validate the JWT token here
    return UserInDB(
        user_id="550e8400-e29b-41d4-a716-446655440000",
        email="test@example.com",
        username="testuser",
        hashed_password="hashed_password_placeholder",
        disabled=False,
        created_at="2023-01-01T00:00:00",
        updated_at="2023-01-01T00:00:00"
    )

# This is a placeholder for database dependency
def get_db():
    # In a real application, this would yield a database session
    # For now, we'll just return None as we're using Beanie's async ODM
    return None
